Based on feedback from folks at enwikinews - having the special page named Special:GoogleNewsSitemap is rather unpopular. Rename to special:NewsFeed.
For now I just made the main en alias for the special page be that. We may
want to rename the extension at some point later.

Also add a version number (Amgine wanted it) Also add my name to the author
list for this extension since I'm feeling a little vain ;)